backend/requirements.txt
Django==4.2.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
dj-database-url==2.2.0
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
